+++ src/main/java/com/example/gimmegonghakauth/InitData.java
@@ -1,14 +1,14 @@
 package com.example.gimmegonghakauth;
 
 import com.example.gimmegonghakauth.constant.AbeekTypeConst;
-import com.example.gimmegonghakauth.dao.AbeekDao;
 import com.example.gimmegonghakauth.dao.MajorsDao;
-import com.example.gimmegonghakauth.user.infrastructure.UserRepository;
-import com.example.gimmegonghakauth.domain.AbeekDomain;
-import com.example.gimmegonghakauth.domain.AbeekDomain.AbeekDomainBuilder;
 import com.example.gimmegonghakauth.domain.MajorsDomain;
+import com.example.gimmegonghakauth.status.domain.AbeekDomain;
+import com.example.gimmegonghakauth.status.domain.AbeekDomain.AbeekDomainBuilder;
+import com.example.gimmegonghakauth.status.domain.MajorName;
+import com.example.gimmegonghakauth.status.infrastructure.AbeekDao;
 import com.example.gimmegonghakauth.user.domain.UserDomain;
-import com.example.gimmegonghakauth.service.recommend.MajorName;
+import com.example.gimmegonghakauth.user.infrastructure.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.context.event.ApplicationReadyEvent;
@@ -41,17 +41,17 @@ public class InitData {
     @Transactional
     public void initDatabase() {
         MajorsDomain computerMajor = MajorsDomain.builder()
-            .id(1L)
-            .major(MajorName.COMPUTER.getName()).build();
+                .id(1L)
+                .major(MajorName.COMPUTER.getName()).build();
         MajorsDomain elecInfoMajor = MajorsDomain.builder()
-            .id(2L)
-            .major(MajorName.ELEC_INFO.getName()).build();
+                .id(2L)
+                .major(MajorName.ELEC_INFO.getName()).build();
         MajorsDomain softwareMajor = MajorsDomain.builder()
-            .id(3L)
-            .major(MajorName.SOFTWARE.getName()).build();
+                .id(3L)
+                .major(MajorName.SOFTWARE.getName()).build();
         MajorsDomain dataScienceMajor = MajorsDomain.builder()
-            .id(4L)
-            .major(MajorName.DATA_SCIENCE.getName()).build();
+                .id(4L)
+                .major(MajorName.DATA_SCIENCE.getName()).build();
         majorsDao.save(computerMajor);
         majorsDao.save(elecInfoMajor);
         majorsDao.save(softwareMajor);
@@ -59,35 +59,35 @@ public class InitData {
 
         //24학년도 computerMajor
         AbeekDomainBuilder abeek1 = AbeekDomain.builder()
-            .abeekType(AbeekTypeConst.BSM)
-            .majorsDomain(computerMajor)
-            .note("this is a test note")
-            .year(24)
-            .minCredit(18);
+                .abeekType(AbeekTypeConst.BSM)
+                .majorsDomain(computerMajor)
+                .note("this is a test note")
+                .year(24)
+                .minCredit(18);
         AbeekDomainBuilder abeek2 = AbeekDomain.builder()
-            .abeekType(AbeekTypeConst.PROFESSIONAL_NON_MAJOR)
-            .majorsDomain(computerMajor)
-            .note("this is a test note")
-            .year(24)
-            .minCredit(14);
+                .abeekType(AbeekTypeConst.PROFESSIONAL_NON_MAJOR)
+                .majorsDomain(computerMajor)
+                .note("this is a test note")
+                .year(24)
+                .minCredit(14);
         AbeekDomainBuilder abeek3 = AbeekDomain.builder()
-            .abeekType(AbeekTypeConst.DESIGN)
-            .majorsDomain(computerMajor)
-            .note("this is a test note")
-            .year(24)
-            .minCredit(10);
+                .abeekType(AbeekTypeConst.DESIGN)
+                .majorsDomain(computerMajor)
+                .note("this is a test note")
+                .year(24)
+                .minCredit(10);
         AbeekDomainBuilder abeek4 = AbeekDomain.builder()
-            .abeekType(AbeekTypeConst.MAJOR)
-            .majorsDomain(computerMajor)
-            .note("this is a test note")
-            .year(24)
-            .minCredit(45);
+                .abeekType(AbeekTypeConst.MAJOR)
+                .majorsDomain(computerMajor)
+                .note("this is a test note")
+                .year(24)
+                .minCredit(45);
         AbeekDomainBuilder abeek5 = AbeekDomain.builder()
-            .abeekType(AbeekTypeConst.MINIMUM_CERTI)
-            .majorsDomain(computerMajor)
-            .note("this is a test note")
-            .year(24)
-            .minCredit(77);
+                .abeekType(AbeekTypeConst.MINIMUM_CERTI)
+                .majorsDomain(computerMajor)
+                .note("this is a test note")
+                .year(24)
+                .minCredit(77);
 
         abeekDao.save(abeek1.build());
         abeekDao.save(abeek2.build());
@@ -97,35 +97,35 @@ public class InitData {
 
         //24학년도 전정통
         AbeekDomainBuilder abeek21 = AbeekDomain.builder()
-            .abeekType(AbeekTypeConst.MSC)
-            .majorsDomain(elecInfoMajor)
-            .note("this is a test note")
-            .year(24)
-            .minCredit(27);
+                .abeekType(AbeekTypeConst.MSC)
+                .majorsDomain(elecInfoMajor)
+                .note("this is a test note")
+                .year(24)
+                .minCredit(27);
         AbeekDomainBuilder abeek22 = AbeekDomain.builder()
-            .abeekType(AbeekTypeConst.PROFESSIONAL_NON_MAJOR)
-            .majorsDomain(elecInfoMajor)
-            .note("this is a test note")
-            .year(24)
-            .minCredit(14);
+                .abeekType(AbeekTypeConst.PROFESSIONAL_NON_MAJOR)
+                .majorsDomain(elecInfoMajor)
+                .note("this is a test note")
+                .year(24)
+                .minCredit(14);
         AbeekDomainBuilder abeek23 = AbeekDomain.builder()
-            .abeekType(AbeekTypeConst.DESIGN)
-            .majorsDomain(elecInfoMajor)
-            .note("this is a test note")
-            .year(24)
-            .minCredit(9);
+                .abeekType(AbeekTypeConst.DESIGN)
+                .majorsDomain(elecInfoMajor)
+                .note("this is a test note")
+                .year(24)
+                .minCredit(9);
         AbeekDomainBuilder abeek24 = AbeekDomain.builder()
-            .abeekType(AbeekTypeConst.MAJOR)
-            .majorsDomain(elecInfoMajor)
-            .note("this is a test note")
-            .year(24)
-            .minCredit(45);
+                .abeekType(AbeekTypeConst.MAJOR)
+                .majorsDomain(elecInfoMajor)
+                .note("this is a test note")
+                .year(24)
+                .minCredit(45);
         AbeekDomainBuilder abeek25 = AbeekDomain.builder()
-            .abeekType(AbeekTypeConst.MINIMUM_CERTI)
-            .majorsDomain(elecInfoMajor)
-            .note("this is a test note")
-            .year(24)
-            .minCredit(86);
+                .abeekType(AbeekTypeConst.MINIMUM_CERTI)
+                .majorsDomain(elecInfoMajor)
+                .note("this is a test note")
+                .year(24)
+                .minCredit(86);
 
         abeekDao.save(abeek21.build());
         abeekDao.save(abeek22.build());
@@ -135,35 +135,35 @@ public class InitData {
 
         //24학년도 소프트웨어
         AbeekDomainBuilder abeek31 = AbeekDomain.builder()
-            .abeekType(AbeekTypeConst.BSM)
-            .majorsDomain(softwareMajor)
-            .note("this is a test note")
-            .year(24)
-            .minCredit(18);
+                .abeekType(AbeekTypeConst.BSM)
+                .majorsDomain(softwareMajor)
+                .note("this is a test note")
+                .year(24)
+                .minCredit(18);
         AbeekDomainBuilder abeek32 = AbeekDomain.builder()
-            .abeekType(AbeekTypeConst.PROFESSIONAL_NON_MAJOR)
-            .majorsDomain(softwareMajor)
-            .note("this is a test note")
-            .year(24)
-            .minCredit(14);
+                .abeekType(AbeekTypeConst.PROFESSIONAL_NON_MAJOR)
+                .majorsDomain(softwareMajor)
+                .note("this is a test note")
+                .year(24)
+                .minCredit(14);
         AbeekDomainBuilder abeek33 = AbeekDomain.builder()
-            .abeekType(AbeekTypeConst.DESIGN)
-            .majorsDomain(softwareMajor)
-            .note("this is a test note")
-            .year(24)
-            .minCredit(12);
+                .abeekType(AbeekTypeConst.DESIGN)
+                .majorsDomain(softwareMajor)
+                .note("this is a test note")
+                .year(24)
+                .minCredit(12);
         AbeekDomainBuilder abeek34 = AbeekDomain.builder()
-            .abeekType(AbeekTypeConst.MAJOR)
-            .majorsDomain(softwareMajor)
-            .note("this is a test note")
-            .year(24)
-            .minCredit(45);
+                .abeekType(AbeekTypeConst.MAJOR)
+                .majorsDomain(softwareMajor)
+                .note("this is a test note")
+                .year(24)
+                .minCredit(45);
         AbeekDomainBuilder abeek35 = AbeekDomain.builder()
-            .abeekType(AbeekTypeConst.MINIMUM_CERTI)
-            .majorsDomain(softwareMajor)
-            .note("this is a test note")
-            .year(24)
-            .minCredit(77);
+                .abeekType(AbeekTypeConst.MINIMUM_CERTI)
+                .majorsDomain(softwareMajor)
+                .note("this is a test note")
+                .year(24)
+                .minCredit(77);
 
         abeekDao.save(abeek31.build());
         abeekDao.save(abeek32.build());
@@ -173,35 +173,35 @@ public class InitData {
 
         //24학년도 데이터사이언스
         AbeekDomainBuilder abeek41 = AbeekDomain.builder()
-            .abeekType(AbeekTypeConst.BSM)
-            .majorsDomain(dataScienceMajor)
-            .note("this is a test note")
-            .year(24)
-            .minCredit(15);
+                .abeekType(AbeekTypeConst.BSM)
+                .majorsDomain(dataScienceMajor)
+                .note("this is a test note")
+                .year(24)
+                .minCredit(15);
         AbeekDomainBuilder abeek42 = AbeekDomain.builder()
-            .abeekType(AbeekTypeConst.PROFESSIONAL_NON_MAJOR)
-            .majorsDomain(dataScienceMajor)
-            .note("this is a test note")
-            .year(24)
-            .minCredit(14);
+                .abeekType(AbeekTypeConst.PROFESSIONAL_NON_MAJOR)
+                .majorsDomain(dataScienceMajor)
+                .note("this is a test note")
+                .year(24)
+                .minCredit(14);
         AbeekDomainBuilder abeek43 = AbeekDomain.builder()
-            .abeekType(AbeekTypeConst.DESIGN)
-            .majorsDomain(dataScienceMajor)
-            .note("this is a test note")
-            .year(24)
-            .minCredit(9);
+                .abeekType(AbeekTypeConst.DESIGN)
+                .majorsDomain(dataScienceMajor)
+                .note("this is a test note")
+                .year(24)
+                .minCredit(9);
         AbeekDomainBuilder abeek44 = AbeekDomain.builder()
-            .abeekType(AbeekTypeConst.MAJOR)
-            .majorsDomain(dataScienceMajor)
-            .note("this is a test note")
-            .year(24)
-            .minCredit(45);
+                .abeekType(AbeekTypeConst.MAJOR)
+                .majorsDomain(dataScienceMajor)
+                .note("this is a test note")
+                .year(24)
+                .minCredit(45);
         AbeekDomainBuilder abeek45 = AbeekDomain.builder()
-            .abeekType(AbeekTypeConst.MINIMUM_CERTI)
-            .majorsDomain(dataScienceMajor)
-            .note("this is a test note")
-            .year(24)
-            .minCredit(74);
+                .abeekType(AbeekTypeConst.MINIMUM_CERTI)
+                .majorsDomain(dataScienceMajor)
+                .note("this is a test note")
+                .year(24)
+                .minCredit(74);
 
         abeekDao.save(abeek41.build());
         abeekDao.save(abeek42.build());
@@ -211,27 +211,27 @@ public class InitData {
 
         //User
         UserDomain user1 = UserDomain.builder()
-            .email("testEmail1@sju.ac.kr")
-            .name("조태현")
-            .password(passwordEncoder.encode("qwer"))
-            .studentId(admin1)
-            .majorsDomain(elecInfoMajor).build();
+                .email("testEmail1@sju.ac.kr")
+                .name("조태현")
+                .password(passwordEncoder.encode("qwer"))
+                .studentId(admin1)
+                .majorsDomain(elecInfoMajor).build();
         userRepository.save(user1);
 
         UserDomain user2 = UserDomain.builder()
-            .email("testEmail2@sju.ac.kr")
-            .name("이희수")
-            .password(passwordEncoder.encode("qwer"))
-            .studentId(admin2)
-            .majorsDomain(dataScienceMajor).build();
+                .email("testEmail2@sju.ac.kr")
+                .name("이희수")
+                .password(passwordEncoder.encode("qwer"))
+                .studentId(admin2)
+                .majorsDomain(dataScienceMajor).build();
         userRepository.save(user2);
 
         UserDomain user3 = UserDomain.builder()
-            .email("testEmail3@sju.ac.kr")
-            .name("홍지섭")
-            .password(passwordEncoder.encode("qwer"))
-            .studentId(admin3)
-            .majorsDomain(computerMajor).build();
+                .email("testEmail3@sju.ac.kr")
+                .name("홍지섭")
+                .password(passwordEncoder.encode("qwer"))
+                .studentId(admin3)
+                .majorsDomain(computerMajor).build();
         userRepository.save(user3);
 
     }
