--- src/main/java/com/example/gimmegonghakauth/dto/GonghakUserDto.java
@@ -1,13 +1,0 @@
-package com.example.gimmegonghakauth.dto;
-
-import java.util.Map;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
-
-@Getter
-@RequiredArgsConstructor
-public class GonghakUserDto {
-
-    private final Map<String, Double> userStatus;
-}
