+++ src/main/java/com/example/gimmegonghakauth/status/domain/AbeekDomain.java
@@ -1,6 +1,7 @@
-package com.example.gimmegonghakauth.domain;
+package com.example.gimmegonghakauth.status.domain;
 
-import com.example.gimmegonghakauth.constant.CourseCategoryConst;
+import com.example.gimmegonghakauth.constant.AbeekTypeConst;
+import com.example.gimmegonghakauth.domain.MajorsDomain;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
@@ -10,6 +11,7 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
+import jakarta.persistence.Lob;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import jakarta.validation.constraints.NotNull;
@@ -21,40 +23,36 @@ import org.hibernate.validator.constraints.Range;
 
 @Entity
 @Data
+@Builder
 @NoArgsConstructor
 @AllArgsConstructor
-@Builder
-@Table(name = "gonghak_course")
-public class GonghakCoursesDomain {
+@Table(name = "abeek")
+public class AbeekDomain {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "gonghak_course_id")
+    @Column(name = "abeek_id")
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "major_id")
     private MajorsDomain majorsDomain;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "course_id")
-    private CoursesDomain coursesDomain;
-
     @NotNull
-    @Range(min = 15, max = 24)
+    @Range(min = 14,max = 24)
     @Column(name = "year")
     private int year;
 
     @NotNull
-    @Enumerated(value = EnumType.STRING)
-    @Column(name = "course_category_const")
-    private CourseCategoryConst courseCategory;
+    @Enumerated(value = EnumType.ORDINAL)
+    @Column(name = "abeek_type")
+    private AbeekTypeConst abeekType;
 
     @NotNull
-    @Column(name = "pass_category")
-    private String passCategory;
+    @Column(name = "min_credit")
+    private int minCredit;
 
-    @NotNull
-    @Range(min = 0,max = 6)
-    @Column(name = "design_credit")
-    private double designCredit;
+    @Lob
+    @Column(name = "note")
+    private String note;
 }
