+++ src/main/java/com/example/gimmegonghakauth/status/infrastructure/GonghakCoursesDao.java
@@ -1,10 +1,10 @@
-package com.example.gimmegonghakauth.dao;
+package com.example.gimmegonghakauth.status.infrastructure;
 
 import com.example.gimmegonghakauth.constant.CourseCategoryConst;
-import com.example.gimmegonghakauth.domain.GonghakCoursesDomain;
+import com.example.gimmegonghakauth.status.domain.GonghakCoursesDomain;
 import com.example.gimmegonghakauth.domain.MajorsDomain;
-import com.example.gimmegonghakauth.dto.GonghakCoursesByMajorDto;
-import com.example.gimmegonghakauth.dto.IncompletedCoursesDto;
+import com.example.gimmegonghakauth.status.service.dto.GonghakCoursesByMajorDto;
+import com.example.gimmegonghakauth.status.service.dto.IncompletedCoursesDto;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
