+++ src/main/java/com/example/gimmegonghakauth/status/service/GonghakCalculateService.java
@@ -1,12 +1,12 @@
-package com.example.gimmegonghakauth.service;
+package com.example.gimmegonghakauth.status.service;
 
 import com.example.gimmegonghakauth.constant.AbeekTypeConst;
-import com.example.gimmegonghakauth.dao.GonghakRepository;
+import com.example.gimmegonghakauth.status.infrastructure.GonghakRepository;
+import com.example.gimmegonghakauth.status.service.dto.GonghakCoursesByMajorDto;
+import com.example.gimmegonghakauth.status.service.dto.GonghakResultDto;
+import com.example.gimmegonghakauth.status.service.dto.GonghakResultDto.ResultPointDto;
+import com.example.gimmegonghakauth.status.service.dto.GonghakStandardDto;
 import com.example.gimmegonghakauth.user.domain.UserDomain;
-import com.example.gimmegonghakauth.dto.GonghakCoursesByMajorDto;
-import com.example.gimmegonghakauth.dto.GonghakResultDto;
-import com.example.gimmegonghakauth.dto.GonghakResultDto.ResultPointDto;
-import com.example.gimmegonghakauth.dto.GonghakStandardDto;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
@@ -27,25 +27,25 @@ public class GonghakCalculateService {
 
         // findStandard -> 학번 입학년도를 기준으로 해당 년도의 abeekType(영역별 구분),minCredit(영역별 인증학점) 불러온다.
         Optional<GonghakStandardDto> standard = gonghakRepository.findStandard(
-            userDomain.getMajorsDomain());
+                userDomain.getMajorsDomain());
 
         // default user abeek 학점 상태 map
         Map<AbeekTypeConst, Double> userAbeekCredit = getUserAbeekCreditDefault(
-            standard.get().getStandards());
+                standard.get().getStandards());
 
         // user 공학 상태 테이블
         // gonghakCourse 중 이수한 과목을 불러온다.
         List<GonghakCoursesByMajorDto> userCoursesByMajorByGonghakCoursesWithCompletedCourses = gonghakRepository.findUserCompletedCourses(
-            userDomain.getStudentId(), userDomain.getMajorsDomain());
+                userDomain.getStudentId(), userDomain.getMajorsDomain());
 
         // user
         // stackUserGonghakCredit -> abeekType에 맞게 이수한 총 학점을 계산한다.
         stackUserGonghakCredit(userCoursesByMajorByGonghakCoursesWithCompletedCourses,
-            userAbeekCredit);
+                userAbeekCredit);
 
         // getUserGonghakResultRatio -> 인증 상태(비율)를 계산한다.
         Map<AbeekTypeConst, ResultPointDto> userResultRatio = getUserGonghakResultRatio(
-            userAbeekCredit, standard);
+                userAbeekCredit, standard);
 
         // 인증 상태(비율) return
         return Optional.of(new GonghakResultDto(userResultRatio));
@@ -53,7 +53,7 @@ public class GonghakCalculateService {
 
     // default user abeek 학점 상태 map을 만들어 반환한다.
     private Map<AbeekTypeConst, Double> getUserAbeekCreditDefault(
-        Map<AbeekTypeConst, Integer> standards) {
+            Map<AbeekTypeConst, Integer> standards) {
         Map<AbeekTypeConst, Double> userAbeekCredit = new ConcurrentHashMap<>();
         Arrays.stream(AbeekTypeConst.values()).forEach(abeekTypeConst -> {
             if (standards.containsKey(abeekTypeConst)) {
@@ -65,34 +65,34 @@ public class GonghakCalculateService {
 
     // 인증 상태(비율)를 계산한다.
     private Map<AbeekTypeConst, ResultPointDto> getUserGonghakResultRatio(
-        Map<AbeekTypeConst, Double> userAbeekCredit,
-        Optional<GonghakStandardDto> standard) {
+            Map<AbeekTypeConst, Double> userAbeekCredit,
+            Optional<GonghakStandardDto> standard) {
 
         Map<AbeekTypeConst, ResultPointDto> userResultRatio = new ConcurrentHashMap<>();
         Arrays.stream(AbeekTypeConst.values()).forEach(abeekTypeConst -> {
-                if (userAbeekCredit.containsKey(abeekTypeConst)) {
-                    getRatio(userAbeekCredit, standard, abeekTypeConst, userResultRatio);
+                    if (userAbeekCredit.containsKey(abeekTypeConst)) {
+                        getRatio(userAbeekCredit, standard, abeekTypeConst, userResultRatio);
+                    }
                 }
-            }
         );
         return userResultRatio;
     }
 
     private void getRatio(Map<AbeekTypeConst, Double> userAbeekCredit,
-        Optional<GonghakStandardDto> standard, AbeekTypeConst abeekTypeConst,
-        Map<AbeekTypeConst, ResultPointDto> userResultRatio) {
+                          Optional<GonghakStandardDto> standard, AbeekTypeConst abeekTypeConst,
+                          Map<AbeekTypeConst, ResultPointDto> userResultRatio) {
 
         userResultRatio.put(
-            abeekTypeConst,
-            new ResultPointDto(userAbeekCredit.get(abeekTypeConst), standard.get().getStandards()
-                .get(abeekTypeConst))
+                abeekTypeConst,
+                new ResultPointDto(userAbeekCredit.get(abeekTypeConst), standard.get().getStandards()
+                        .get(abeekTypeConst))
         );
     }
 
     // abeekType에 맞게 이수한 총 학점을 계산한다.
     private void stackUserGonghakCredit(
-        List<GonghakCoursesByMajorDto> userCoursesByMajorByGonghakCoursesWithCompletedCourses,
-        Map<AbeekTypeConst, Double> userAbeekCredit) {
+            List<GonghakCoursesByMajorDto> userCoursesByMajorByGonghakCoursesWithCompletedCourses,
+            Map<AbeekTypeConst, Double> userAbeekCredit) {
         userCoursesByMajorByGonghakCoursesWithCompletedCourses.forEach(gonghakCoursesByMajorDto -> {
             switch (gonghakCoursesByMajorDto.getCourseCategory()) {
                 case 전공:
@@ -100,11 +100,11 @@ public class GonghakCalculateService {
                     break;
                 case 전문교양:
                     stackCredit(AbeekTypeConst.PROFESSIONAL_NON_MAJOR, gonghakCoursesByMajorDto,
-                        userAbeekCredit);
+                            userAbeekCredit);
                     break;
                 case 교양:
                     stackCredit(AbeekTypeConst.NON_MAJOR, gonghakCoursesByMajorDto,
-                        userAbeekCredit);
+                            userAbeekCredit);
                     break;
                 case MSC:
                     stackCredit(AbeekTypeConst.MSC, gonghakCoursesByMajorDto, userAbeekCredit);
@@ -120,15 +120,15 @@ public class GonghakCalculateService {
 
 
     private void stackCredit(AbeekTypeConst abeekTypeConst,
-        GonghakCoursesByMajorDto gonghakCoursesByMajorDto,
-        Map<AbeekTypeConst, Double> userAbeekCredit) {
+                             GonghakCoursesByMajorDto gonghakCoursesByMajorDto,
+                             Map<AbeekTypeConst, Double> userAbeekCredit) {
         double inputCredit = getInputCredit(abeekTypeConst, gonghakCoursesByMajorDto);
         userAbeekCredit.put(abeekTypeConst, userAbeekCredit.get(abeekTypeConst) + inputCredit);
 
     }
 
     private double getInputCredit(AbeekTypeConst abeekTypeConst,
-        GonghakCoursesByMajorDto gonghakCoursesByMajorDto) {
+                                  GonghakCoursesByMajorDto gonghakCoursesByMajorDto) {
         if (abeekTypeConst == AbeekTypeConst.DESIGN) {
             return gonghakCoursesByMajorDto.getDesignCredit();
         } else {
