+++ src/main/java/com/example/gimmegonghakauth/status/service/dto/GonghakCoursesByMajorDto.java
@@ -1,15 +1,17 @@
-package com.example.gimmegonghakauth.dto;
+package com.example.gimmegonghakauth.status.service.dto;
 
-import com.example.gimmegonghakauth.constant.AbeekTypeConst;
 import com.example.gimmegonghakauth.constant.CourseCategoryConst;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public class IncompletedCoursesDto {
+public class GonghakCoursesByMajorDto {
+    private final Long courseId;
     private final String courseName;
+    private final int year;
     private final CourseCategoryConst courseCategory;
-    private final int credit;
+    private final String passCategory;
     private final double designCredit;
+    private final int credit;
 }
