+++ src/main/java/com/example/gimmegonghakauth/status/service/dto/GonghakRecommendCoursesDto.java
@@ -1,7 +1,6 @@
-package com.example.gimmegonghakauth.dto;
+package com.example.gimmegonghakauth.status.service.dto;
 
 import com.example.gimmegonghakauth.constant.AbeekTypeConst;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
