+++ src/main/java/com/example/gimmegonghakauth/status/service/dto/GonghakResultDto.java
@@ -1,4 +1,4 @@
-package com.example.gimmegonghakauth.dto;
+package com.example.gimmegonghakauth.status.service.dto;
 
 import com.example.gimmegonghakauth.constant.AbeekTypeConst;
 import java.util.Map;
