+++ src/main/java/com/example/gimmegonghakauth/status/service/dto/GonghakStandardDto.java
@@ -1,11 +1,10 @@
-package com.example.gimmegonghakauth.dto;
+package com.example.gimmegonghakauth.status.service.dto;
 
 
 import com.example.gimmegonghakauth.constant.AbeekTypeConst;
 import java.util.Map;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
 
 @Getter
 @RequiredArgsConstructor
