+++ src/main/java/com/example/gimmegonghakauth/status/service/dto/IncompletedCoursesDto.java
@@ -1,18 +1,14 @@
-package com.example.gimmegonghakauth.dto;
+package com.example.gimmegonghakauth.status.service.dto;
 
 import com.example.gimmegonghakauth.constant.CourseCategoryConst;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
 
 @Getter
 @RequiredArgsConstructor
-public class GonghakCoursesByMajorDto {
-    private final Long courseId;
+public class IncompletedCoursesDto {
     private final String courseName;
-    private final int year;
     private final CourseCategoryConst courseCategory;
-    private final String passCategory;
-    private final double designCredit;
     private final int credit;
+    private final double designCredit;
 }
