+++ src/main/java/com/example/gimmegonghakauth/status/service/recommend/GonghakRecommendService.java
@@ -1,7 +1,7 @@
-package com.example.gimmegonghakauth.service.recommend;
+package com.example.gimmegonghakauth.status.service.recommend;
 
+import com.example.gimmegonghakauth.status.service.dto.GonghakRecommendCoursesDto;
 import com.example.gimmegonghakauth.user.domain.UserDomain;
-import com.example.gimmegonghakauth.dto.GonghakRecommendCoursesDto;
 
 public interface GonghakRecommendService {
     GonghakRecommendCoursesDto createRecommendCourses(UserDomain userDomain);
