+++ src/main/java/com/example/gimmegonghakauth/status/service/recommend/RecommendServiceSelectManager.java
@@ -1,9 +1,9 @@
-package com.example.gimmegonghakauth.service.recommend;
+package com.example.gimmegonghakauth.status.service.recommend;
 
-import static com.example.gimmegonghakauth.service.recommend.MajorName.*;
+import static com.example.gimmegonghakauth.status.domain.MajorName.ELEC_INFO;
 
-import com.example.gimmegonghakauth.user.infrastructure.UserRepository;
 import com.example.gimmegonghakauth.domain.MajorsDomain;
+import com.example.gimmegonghakauth.user.infrastructure.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationContext;
 import org.springframework.stereotype.Component;
@@ -20,12 +20,12 @@ public class RecommendServiceSelectManager {
     // 학과에 따른 추천 서비스를 설정한다.
     public GonghakRecommendService selectRecommendService(Long studentId) {
         MajorsDomain majorsDomain = userRepository.findByStudentId(studentId).get()
-            .getMajorsDomain();
+                .getMajorsDomain();
         if (majorsDomain.getMajor().contains(ELEC_INFO.getName())) {
             return applicationContext.getBean("elecInfoMajorGonghakRecommendService",
-                ElecInfoMajorGonghakRecommendService.class);
+                    ElecInfoMajorGonghakRecommendService.class);
         }
         return applicationContext.getBean("computerMajorGonghakRecommendService",
-            ComputerMajorGonghakRecommendService.class);
+                ComputerMajorGonghakRecommendService.class);
     }
 }
