+++ src/test/java/com/example/gimmegonghakauth/Service/GonghakCalculateServiceTest.java
@@ -4,14 +4,14 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import com.example.gimmegonghakauth.Service.GonghakCalculateServiceTest.CalculateTestConfig;
 import com.example.gimmegonghakauth.constant.AbeekTypeConst;
-import com.example.gimmegonghakauth.dao.AbeekDao;
-import com.example.gimmegonghakauth.dao.GonghakCoursesDao;
-import com.example.gimmegonghakauth.dao.GonghakDao;
-import com.example.gimmegonghakauth.dao.GonghakRepository;
 import com.example.gimmegonghakauth.domain.MajorsDomain;
+import com.example.gimmegonghakauth.status.infrastructure.AbeekDao;
+import com.example.gimmegonghakauth.status.infrastructure.GonghakCoursesDao;
+import com.example.gimmegonghakauth.status.infrastructure.GonghakDao;
+import com.example.gimmegonghakauth.status.infrastructure.GonghakRepository;
+import com.example.gimmegonghakauth.status.service.GonghakCalculateService;
+import com.example.gimmegonghakauth.status.service.dto.GonghakResultDto.ResultPointDto;
 import com.example.gimmegonghakauth.user.domain.UserDomain;
-import com.example.gimmegonghakauth.dto.GonghakResultDto.ResultPointDto;
-import com.example.gimmegonghakauth.service.GonghakCalculateService;
 import java.util.Map;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -37,58 +37,58 @@ class GonghakCalculateServiceTest {
     private GonghakCalculateService gonghakCalculateService;
 
     private static final MajorsDomain TEST_MAJORSDOMAIN = MajorsDomain.builder()
-        .id(1L)
-        .major("컴퓨터공학과").build();
+            .id(1L)
+            .major("컴퓨터공학과").build();
 
     private static final UserDomain TEST_USERDOMAIN = UserDomain.builder()
-        .email("testEmail")
-        .name("홍지섭")
-        .password("qwer")
-        .studentId(19011706L)
-        .majorsDomain(TEST_MAJORSDOMAIN).build();
+            .email("testEmail")
+            .name("홍지섭")
+            .password("qwer")
+            .studentId(19011706L)
+            .majorsDomain(TEST_MAJORSDOMAIN).build();
 
     @TestConfiguration
     @RequiredArgsConstructor
-    static class CalculateTestConfig{
+    static class CalculateTestConfig {
         private final AbeekDao abeekDao;
         private final GonghakCoursesDao gonghakCoursesDao;
 
         @Bean
-        public GonghakRepository gonghakRepository(){
+        public GonghakRepository gonghakRepository() {
             return new GonghakDao(abeekDao, gonghakCoursesDao);
         }
 
         @Bean
-        public GonghakCalculateService gonghakCalculateService(){
+        public GonghakCalculateService gonghakCalculateService() {
             return new GonghakCalculateService(gonghakRepository());
         }
     }
 
     @Test
     @DisplayName("check log gonghakCalculateServiceTest")
-    void logGonghakCalculateServiceTest(){
-            Map<AbeekTypeConst, ResultPointDto> userResultRatio = gonghakCalculateService.getResultRatio(
-            TEST_USERDOMAIN).get().getUserResultRatio();
+    void logGonghakCalculateServiceTest() {
+        Map<AbeekTypeConst, ResultPointDto> userResultRatio = gonghakCalculateService.getResultRatio(
+                TEST_USERDOMAIN).get().getUserResultRatio();
 
-        log.info("userResultRatio = {}",userResultRatio);
+        log.info("userResultRatio = {}", userResultRatio);
     }
 
     @Test
     @DisplayName("컴퓨터공학과 GonghakCalculateService 계산 결과 체크")
-    void correctGonghakCalculateServiceTestCom(){
+    void correctGonghakCalculateServiceTestCom() {
         Map<AbeekTypeConst, ResultPointDto> userResultRatio = gonghakCalculateService.getResultRatio(
-            TEST_USERDOMAIN).get().getUserResultRatio();
+                TEST_USERDOMAIN).get().getUserResultRatio();
 
         userResultRatio.keySet().forEach(
-            abeekTypeConst -> {
-                Double userPoint = userResultRatio.get(abeekTypeConst).getUserPoint();
-                assertThat(userPoint).isLessThanOrEqualTo(1);
-                assertThat(userPoint).isGreaterThanOrEqualTo(0);
-            }
+                abeekTypeConst -> {
+                    Double userPoint = userResultRatio.get(abeekTypeConst).getUserPoint();
+                    assertThat(userPoint).isLessThanOrEqualTo(1);
+                    assertThat(userPoint).isGreaterThanOrEqualTo(0);
+                }
         );
     }
 
-    private double getExpectedCredit(int credit,Map<AbeekTypeConst, Integer> standard,AbeekTypeConst abeekTypeConst) {
-        return Double.valueOf(String.format("%.4f",(double) credit / standard.get(abeekTypeConst)));
+    private double getExpectedCredit(int credit, Map<AbeekTypeConst, Integer> standard, AbeekTypeConst abeekTypeConst) {
+        return Double.valueOf(String.format("%.4f", (double) credit / standard.get(abeekTypeConst)));
     }
 }
