+++ src/test/java/com/example/gimmegonghakauth/Service/GonghakRecommendServiceTest.java
@@ -3,13 +3,13 @@ package com.example.gimmegonghakauth.Service;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.example.gimmegonghakauth.constant.AbeekTypeConst;
-import com.example.gimmegonghakauth.user.infrastructure.UserRepository;
-import com.example.gimmegonghakauth.user.domain.UserDomain;
-import com.example.gimmegonghakauth.dto.IncompletedCoursesDto;
 import com.example.gimmegonghakauth.dto.LoginDto;
-import com.example.gimmegonghakauth.service.recommend.ComputerMajorGonghakRecommendService;
-import com.example.gimmegonghakauth.service.recommend.GonghakRecommendService;
-import com.example.gimmegonghakauth.service.recommend.RecommendServiceSelectManager;
+import com.example.gimmegonghakauth.status.service.dto.IncompletedCoursesDto;
+import com.example.gimmegonghakauth.status.service.recommend.ComputerMajorGonghakRecommendService;
+import com.example.gimmegonghakauth.status.service.recommend.GonghakRecommendService;
+import com.example.gimmegonghakauth.status.service.recommend.RecommendServiceSelectManager;
+import com.example.gimmegonghakauth.user.domain.UserDomain;
+import com.example.gimmegonghakauth.user.infrastructure.UserRepository;
 import java.util.List;
 import java.util.Map;
 import lombok.extern.slf4j.Slf4j;
@@ -36,44 +36,45 @@ class GonghakRecommendServiceTest {
 
     @Test
     @DisplayName("유저의 학과에 맞는 recommendService 클래스 가져오기 확인 - 컴퓨터공학과")
-    void recommendServiceSelectManagerTest(){
+    void recommendServiceSelectManagerTest() {
         LoginDto loginDto = new LoginDto(COM_TEST_STUDENT_ID);
         GonghakRecommendService gonghakRecommendService = recommendServiceSelectManager.selectRecommendService(
-            loginDto.getStudentId());
+                loginDto.getStudentId());
         assertThat(gonghakRecommendService).isInstanceOf(
-            ComputerMajorGonghakRecommendService.class);
+                ComputerMajorGonghakRecommendService.class);
     }
 
     @Test
-    void createRecommendCoursesTest(){
+    void createRecommendCoursesTest() {
         UserDomain testUser = userRepository.findByStudentId(COM_TEST_STUDENT_ID).get();
 
         GonghakRecommendService comGonghakRecommendService = recommendServiceSelectManager.selectRecommendService(
-            new LoginDto(COM_TEST_STUDENT_ID).getStudentId());
+                new LoginDto(COM_TEST_STUDENT_ID).getStudentId());
         Map<AbeekTypeConst, List<IncompletedCoursesDto>> recommendCoursesByAbeekType = comGonghakRecommendService.createRecommendCourses(
-            testUser).getRecommendCoursesByAbeekType();
+                testUser).getRecommendCoursesByAbeekType();
 
-        log.info("recommendCoursesByAbeekType.keySet()= {}",recommendCoursesByAbeekType.keySet());
+        log.info("recommendCoursesByAbeekType.keySet()= {}", recommendCoursesByAbeekType.keySet());
         //[PROFESSIONAL_NON_MAJOR, BSM, DESIGN, MAJOR, MINIMUM_CERTI]
 
         recommendCoursesByAbeekType.keySet().forEach(
-            abeekTypeConst -> {
-                log.info("abeekTypeConst.name() = {}",abeekTypeConst.name());
-                recommendCoursesByAbeekType.get(abeekTypeConst);
-                for (IncompletedCoursesDto incompletedCoursesDto : recommendCoursesByAbeekType.get(abeekTypeConst)) {
-                    log.info("incompletedCoursesDto.getCourseName= {}",incompletedCoursesDto.getCourseName());
+                abeekTypeConst -> {
+                    log.info("abeekTypeConst.name() = {}", abeekTypeConst.name());
+                    recommendCoursesByAbeekType.get(abeekTypeConst);
+                    for (IncompletedCoursesDto incompletedCoursesDto : recommendCoursesByAbeekType.get(
+                            abeekTypeConst)) {
+                        log.info("incompletedCoursesDto.getCourseName= {}", incompletedCoursesDto.getCourseName());
+                    }
                 }
-            }
         );
 
         assertThat(recommendCoursesByAbeekType.get(AbeekTypeConst.PROFESSIONAL_NON_MAJOR).size()).isNotEqualTo(0);
 
         assertThat(recommendCoursesByAbeekType.keySet()).containsOnly(
-            AbeekTypeConst.PROFESSIONAL_NON_MAJOR,
-            AbeekTypeConst.BSM,
-            AbeekTypeConst.DESIGN,
-            AbeekTypeConst.MAJOR,
-            AbeekTypeConst.MINIMUM_CERTI
+                AbeekTypeConst.PROFESSIONAL_NON_MAJOR,
+                AbeekTypeConst.BSM,
+                AbeekTypeConst.DESIGN,
+                AbeekTypeConst.MAJOR,
+                AbeekTypeConst.MINIMUM_CERTI
         );
     }
 }
